+++ apps/rig_auth/lib/rig_auth/authorization_check/external.ex
@@ -8,15 +8,13 @@ defmodule RigAuth.AuthorizationCheck.External do
 
   @json_mimetype "application/json; charset=utf-8"
 
-  @spec check(url :: String.t(), req_headers :: Conn.headers(), body_params :: map) ::
+  @spec check(url :: String.t(), req_headers :: Conn.headers(), body :: binary) ::
           true | false | {:error, url :: String.t(), error :: any()}
-  def check(url, req_headers, body_params) do
+  def check(url, req_headers, body) do
     headers =
       for {"authorization", _} = header <- req_headers,
           into: %{"content-type" => @json_mimetype},
           do: header
-
-    body = Poison.encode!(body_params)
 
     case HTTPoison.post(url, body, headers) do
       {:ok, %HTTPoison.Response{status_code: status}} when status >= 200 and status < 300 -> true
@@ -26,10 +24,10 @@ defmodule RigAuth.AuthorizationCheck.External do
     end
   end
 
-  @spec check_or_log(base_url :: String.t(), req_headers :: Conn.headers(), body_params :: map) ::
+  @spec check_or_log(base_url :: String.t(), req_headers :: Conn.headers(), body :: binary) ::
           :ok | {:error, :not_authorized}
-  def check_or_log(base_url, req_headers, body_params) do
-    case check(base_url, req_headers, body_params) do
+  def check_or_log(base_url, req_headers, body) do
+    case check(base_url, req_headers, body) do
       true ->
         :ok