+++ apps/rig_inbound_gateway/lib/rig_inbound_gateway/events.ex
@@ -2,36 +2,41 @@ defmodule RigInboundGateway.Events do
   @moduledoc """
   Utility functions used in more than one controller.
   """
-  alias CloudEvent
+  alias UUID
+
   alias Rig.Connection
   alias Rig.Subscription
+  alias RigCloudEvents.CloudEvent
 
   @spec welcome_event() :: CloudEvent.t()
   def welcome_event do
     connection_pid = self()
     connection_token = Connection.Codec.serialize(connection_pid)
-    data = %{connection_token: connection_token}
 
-    CloudEvent.new!(%{
-      "cloudEventsVersion" => "0.1",
-      "eventType" => "rig.connection.create",
-      "source" => "rig"
-    })
-    |> CloudEvent.with_data(data)
+    rig_event(
+      "rig.connection.create",
+      %{connection_token: connection_token}
+    )
   end
 
   @spec subscriptions_set([Subscription.t()]) :: CloudEvent.t()
   def subscriptions_set(subscriptions) do
-    CloudEvent.new!(%{
-      "cloudEventsVersion" => "0.1",
-      "eventType" => "rig.subscriptions_set",
-      "source" => "rig"
-    })
-    |> CloudEvent.with_data(
-      subscriptions
-      |> Enum.map(fn %Subscription{event_type: event_type, constraints: constraints} ->
+    rig_event(
+      "rig.subscriptions_set",
+      Enum.map(subscriptions, fn %Subscription{event_type: event_type, constraints: constraints} ->
         %{"eventType" => event_type, "oneOf" => constraints}
       end)
     )
+  end
+
+  defp rig_event(type, data) do
+    CloudEvent.parse!(%{
+      specversion: "0.2",
+      type: type,
+      source: "rig",
+      id: UUID.uuid4(),
+      time: Timex.now() |> Timex.format!("{RFC3339}"),
+      data: data
+    })
   end
 end